Modularize native library
