+++ src/native/str.rs
@@ -1,0 +1,5 @@
+use crate::eval::{Environment};
+
+pub fn add_str_procs(_env: &mut Environment) {
+	// TODO
+}
